findpeaks: drop usage of inputParser to avoid dependency on general package.
* findpeaks.m: both the general package and the upcoming Octave 4.0 release
have an implementation of inputParser.  This causes issues so they can't
co-exist.  While our current solution means it works in both Octave 3.8 and
4.0, it is still depedent on the general package.  General is about to be
release either without inputParser or depending on (octave < 4.0.0) which
will cause issues.  So we just drop the depedency.  Input checking with
inputParser is simpler but at the moment is not worth the extra depedency.
* DESCRIPTION: drop dependecy on general.
* NEWS: note reduced dependency.